Rewrite logSDLRendererInfo()
The function logSDLRendererInfo() works but can use rewriting.
Add preliminary `"Renderer infos"` message with initial newline, and concluding
newline to better delimitate the messages block that constitutes the logging of
the renderer infos.
Add `0x` prefixes to number output in hexadecimal.
Change the names of the variables: *reninfo* -> *i* and *i* -> *k*.
Change the memory width of variables: *n* and *i* are now `char`s rather than `int`s
(since they are certain to be <= 16).