Change the type of a member, modify some functions.
The member *geom* is now a *Rect* instead of *SDL_Rect* (those are the same type, but
it is be for coherence to use the same all through the project).
The function *updatesprite()* does not use a cast to *SDL_Rect* anymore, it uses the
*rect()* function defined in `geom.h` (for the exact same reaseon as above).
Remove a commented line in that function too.

The functions *isptinsprite()* now uses the function *isptinrect* from `geom.h`.
Note: *isptinrect()* takes two pointers as arguments, but *isptinsprite()* takes a plain
*Point* object.