Totally reworked the text.h module.
The module now provides a type, *Text*, which holds textual data and font and color to
display in, an (optional) background color and a type of rendering to apply to the text.

Typical functions are implemented (*newtext()*, *freetext()*, *updatetext()*, setters)
and most importantly *drawtext()*, which does the actual text rendering on screen.
(Though the window has to been *rendered* afterwards for it to be displayed, see previous
commit.)

Complete documentation is included.