Move game loop as a function of Window.
The main loop is the block that orchestrates the running of the game.
It is now moved as the *windowRun()* function, executing a function
specified in *setWindowGameLoop()*. Its profile is:

    void theGameLoopFunction(void);

Though it has no arguments, it may (and most generally will) interact
with global variables, such as the window itelf, sprites, or event
handlers.