Remove usages of size_t type
All occurrences of `size_t` are replaced with `unsigned int`. This type
is shorter in size (`int` and not `long`) and does not depend on an
external header (`stdlib.h`). The #include for this header is moved from
the `scene.h` header to its source to still provide *malloc* and *free*.