Replace wildcard with variable
The object files were mentioned with a wildcard as `$(OBJ_DIR)/*.o`,
which can be buggy in case when some of the files must not be used (e.g.
tests suites) Though this is not the case now, the mention is changed
to a mere mention of `$(OBJ)` which is assured to contain *only* the
object files needed for main build.