Add GUI elements
The new module *uilements.h* introduces functions to create interface
elements. These elements are merely sprites, but their interpretation
is different, as they are not elements of the game.

There are (for now) the functions *label()* and *button()*; the former
creating a non-interactible element whose purpose is only to display
tetual information; the latter creating an element having the ability
to be interacted with by the player -- through a mouse click -- to
perform a specific action.

The module also introduces a set of functions to set up a GUI style:
text font and color, background color (for the buttons).