Two types of scenes: game scene, UI scene
A boolean is added to the Scene structure (the member is *isui*) to
differentiate whether the scene is an UI scene (ie. a menu) or a game
scene.
A game scene contains sprites, a UI scene contains *UIElement*s; the
latter is a new type defined in `uielements.h` because their behaviour
differs too much from a Sprite to be implemented as one.

The *sprites* members is thus renamed *elements*, and contains *Sprite*s
if the scene is a game one, and *UIElement*s if it is a UI scene --
it is left to the programmer to ensure that *all* items added are of the
right type -- and the *Scene* methods are duplicated according to these
different types, and calling a game scene method with a UI scene will
return a `false`-y value (ie. `false`, `0`, or `NULL`, according to the
function's return type).