Change way of handling events
An *EventHandler* is now a function specific to a type of event (be it
a mouse motion, keyboard hit...) which receives as argument the event to
handle, in its specific type (not the generic union *SDL_Event*).

Concerning the mouse event types, the former argument (a *Point*) is
retrievable through the fields *x* and *y* of the event.

The handler functions (e.g. *onMouseMove()*) do not return the handler
anymore, but call it themselves with the event given as argument.
This allows to lighten the *main()* function, and it will make possible
to implement a mouse click directly from the events module.