Handle mouse clicks
A mouse click is the event triggered when a mouse down is immediately
followed by a mouse release.
The triggering is made right after the call to the mouse up handler
if the position of the event is the same (or approximately the same)
than the position of the last mouse down event.
On the side of the user, the handling is defined is the same way that
mouse down or mouse up.

The setting of a mouse button handler is refactored in a `static`
function named *setHandler()*; this functio is used in
*set_onMouseDown()*, *set_onMouseUp()* and *set_onMouseClick()*.