Change behaviour of in-game sprites
Sprites on the scene can no longer be dragged along, but are only
pickable (when clicked, go to the inventory).
Sprites in the inventory can be dragged to the scene. However, when
released on the scene, they belong to the scene again, so they must be
picked up once more to be able to be dragged on the scene somewhere
else.

A mouse click handler is implemented for the left mouse button, the
`static` function *leftclick()*. This is the function that manages the
pick-up of the sprites from the scene, no longer *leftup()*.
As for the latter, itself, *leftdown()* and *move()* handle the drag
and drop of a sprite from the inventory to the scene.