Shrink group of colors #defines
The brief descriptions are moved inline after the definition of each color
constant, the empty lines between each them are removed.

NB: an `enum` *cannot* be used because enum values must be compatible with
the `int` type, and e.g. *YELLOW* is 32-bits wide and the first bit set,
this value would only fit on an `unsigned int`.