Fix variables and compiler options
The common variable `LDFLAGS` was misused in the compilation rule
instead of linkage. This is fixed, and the variable is now declared
empty (kept for future uses). Its content is moved to the also common
variable `CFLAGS`, used in the compilation step.

Consequelnty `LDFLAGS` is no longer mentioned in the compilation rule,
and is mentioned instead of `CFLAGS` in the linking recipe.